Add files via upload
Write an application that prompts a user to enter a letter and creates a thread that prints that letter 50 times. after 25 prints of the same letter, another thread should be created to print 25 times the letter that comes just after that letter in alphabetical order.
Eg. If a was entered, a second threads will be created to print b 25 times.